Implementacion de Assets , Rampas ,  buggs arreglados y Sound Manager
He implementado el asset del sniper y la metralladora he hecho que cuando este encima de una rampa se aplique la friccion y he arreglado un bugg de cuando cogias el sniper se buggeaba y he hecho que el sound manager guarde el audio correctamente

+++ HELLBOTS/Assets/Scripts/Managers/SoundManager.cs
@@ -24,8 +24,8 @@ public class SoundManager : MonoBehaviour
         if (File.Exists("sound.sav"))
         {
             reader = new BinaryReader(File.Open("sound.sav", FileMode.Open));
-            Music = reader.ReadInt32();
-            SFX = reader.ReadInt32();
+            Music = reader.ReadSingle();
+            SFX = reader.ReadSingle();
             reader.Close();
         }
         else { return; }
